+++ src/i18n.cpp
@@ -140,10 +140,10 @@ bool Catalogue::load_resource(const string& resource_name)
         length_offset_origonal.resize(header.count * 2);
         length_offset_translated.resize(header.count * 2);
         stream->seek(header.offset_origonal);
-        if (stream->read(length_offset_origonal.data(), length_offset_origonal.size() * sizeof(uint32_t)) != int(length_offset_origonal.size() * sizeof(uint32_t)))
+        if (stream->read(length_offset_origonal.data(), length_offset_origonal.size() * sizeof(uint32_t)) != length_offset_origonal.size() * sizeof(uint32_t))
             return false;
         stream->seek(header.offset_translated);
-        if (stream->read(length_offset_translated.data(), length_offset_translated.size() * sizeof(uint32_t)) != int(length_offset_translated.size() * sizeof(uint32_t)))
+        if (stream->read(length_offset_translated.data(), length_offset_translated.size() * sizeof(uint32_t)) != length_offset_translated.size() * sizeof(uint32_t))
             return false;
         if (swap)
         {
